+++ src/controllers/modificationRequest/index.ts
@@ -4,7 +4,7 @@ export * from './getModificationRequestListForPorteurPage'
 export * from './getModificationRequestPage'
 export * from './postAccorderOuRejeterDemandeDelai'
 export * from './postAnnulerDemandeDélai'
-export * from './postAnnulerRejetDemandeDélai'
+export * from './postAnnulerRejetDélai'
 export * from './postCancelModificationRequest'
 export * from './postConfirmRequest'
 export * from './postDemanderDélai'
