+++ src/infra/sequelize/projections/modificationRequest/updates/onRejetDélaiAnnulé.ts
@@ -1,9 +1,9 @@
 import { logger } from '@core/utils'
-import { RejetDemandeDélaiAnnulé } from '@modules/demandeModification'
+import { RejetDélaiAnnulé } from '@modules/demandeModification'
 
-export const onRejetDemandeDélaiAnnulé =
+export const onRejetDélaiAnnulé =
   (models) =>
-  async ({ payload }: RejetDemandeDélaiAnnulé) => {
+  async ({ payload }: RejetDélaiAnnulé) => {
     const { demandeDélaiId } = payload
     try {
       const ModificationRequestModel = models.ModificationRequest
