+++ src/infra/sequelize/projectionsNext/projectEvents/updates/index.ts
@@ -42,4 +42,4 @@ export * from './onProjectNotified'
 export * from './onProjectPTFRemoved'
 export * from './onProjectPTFSubmitted'
 export * from './onProjectStepStatusUpdated'
-export * from './onRejetDemandeDélaiAnnulé'
+export * from './onRejetDélaiAnnulé'
