--- src/modules/demandeModification/demandeDélai/annulerRéponse/index.ts
@@ -1,1 +1,0 @@
-export * from './annulerRéponseDemandeDélai'
