+++ src/modules/demandeModification/demandeDélai/events/index.ts
@@ -3,4 +3,4 @@ export * from './DélaiAnnulé'
 export * from './DélaiDemandé'
 export * from './DélaiEnInstruction'
 export * from './DélaiRejeté'
-export * from './RejetDemandeDélaiAnnulé'
+export * from './RejetDélaiAnnulé'
