+++ src/modules/demandeModification/demandeDélai/index.ts
@@ -1,6 +1,6 @@
 export * from './accorder'
 export * from './annuler'
-export * from './annulerRéponse'
+export * from './annulerRejet'
 export * from './demander'
 export * from './events'
 export * from './passerEnInstruction'
