+++ src/modules/notification/eventHandlers/demandeDélai/index.ts
@@ -3,4 +3,4 @@ export * from './onDélaiAnnulé'
 export * from './onDélaiDemandé'
 export * from './onDélaiEnInstruction'
 export * from './onDélaiRejeté'
-export * from './onRejetDemandeDélaiAnnulé'
+export * from './onRejetDélaiAnnulé'
